+++ packages/jest-cli/src/init/generate_config_file.ts
@@ -35,7 +35,13 @@ const generateConfigFile = (
   results: Record<string, unknown>,
   generateEsm = false,
 ): string => {
-  const {coverage, coverageProvider, clearMocks, environment} = results;
+  const {
+    useTypescript,
+    coverage,
+    coverageProvider,
+    clearMocks,
+    environment,
+  } = results;
 
   const overrides: Record<string, unknown> = {};
 
@@ -81,10 +87,20 @@ const generateConfigFile = (
     }
   }
 
+  const configHeaderMessage = `/*
+ * For a detailed explanation regarding each configuration property${
+   useTypescript ? ' and type check' : ''
+ }, visit:
+ * https://jestjs.io/docs/en/configuration.html
+ */
+
+`;
+
   return (
-    '// For a detailed explanation regarding each configuration property, visit:\n' +
-    '// https://jestjs.io/docs/en/configuration.html\n\n' +
-    (generateEsm ? 'export default {\n' : 'module.exports = {\n') +
+    configHeaderMessage +
+    (useTypescript || generateEsm
+      ? 'export default {\n'
+      : 'module.exports = {\n') +
     properties.join('\n') +
     '};\n'
   );